Update main.c
me falto castiar

+++ P1/P1E2/Src/main.c
@@ -40,7 +40,7 @@ int main(void)
   BSP_PB_Init(BUTTON_USER, BUTTON_MODE_GPIO);
 
   uint8_t secuencia[3] = {LED1,LED2,LED3}; //Los leds estan representados y ordenados en un enum, por ende los puedo trabajar con numeros (LED1 = 0, LED2= 1, LED3 = 3)
-  int8_t i = 0;
+  int8_t i = 0; 
   uint8_t ordenDeSecuencia = 0; //  0 = secuencia 1;   1 = secuencia 2
 
   while (1)
@@ -75,8 +75,8 @@ int main(void)
 
 	//chequeo si i es mayor al tamaño del array (Limitadores)
 
-	if(i >= sizeof(secuencia)/sizeof(secuencia[0])) { i=0;}
-	if(i<0){ i = (sizeof(secuencia)/sizeof(secuencia[0]))-1;}
+	if(i >= (uint8_t)(sizeof(secuencia)/sizeof(secuencia[0]))) { i=0;}
+	if(i<0){ i = (uint8_t)((sizeof(secuencia)/sizeof(secuencia[0]))-1);}
 
   }
 }
